+++ packages/crypto-commit/test/fixtures/block.ts
@@ -1,7 +1,5 @@
 import type { Contracts } from "@mainsail/contracts";
 
-import { BigNumber } from "@mainsail/utils";
-
 export const validatorMnemonic =
 	"sudden head royal retire duck discover danger then basic rice wish left whip chronic enrich sun behind idea remind retire coyote select goddess exile";
